+++ pkgs/intl_translation/lib/src/messages/submessages/submessage.dart
@@ -95,7 +95,14 @@ abstract class SubMessage extends ComplexMessage {
   List toJson() {
     var json = [];
     json.add(dartMessageName);
-    json.add(arguments.indexOf(mainArgument));
+    var indexOf = arguments.indexOf(mainArgument);
+    if (indexOf == -1) {
+      // This is an error as this should have been checked in
+      // lib/visitors/plural_gender_visitor.dart already.
+      throw ArgumentError('The argument `$mainArgument` could not be found in '
+          '`$arguments`.');
+    }
+    json.add(indexOf);
     for (var arg in codeAttributeNames) {
       json.add(this[arg]?.toJson());
     }