+++ pkgs/intl_translation/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.21.0-wip
   * BREAKING CHANGE: Update `dart_style` to `^3.0.0`
   * Allow analyzer `>=6.3.0 <8.0.0`
+  * Throw if the main argument in plurals and selects does not match the argument list.
 
 ## 0.20.1
   * Add topics to `pubspec.yaml`